requirements.txt
click==8.0.3
colorama==0.4.4
dnspython==2.1.0
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pymongo==4.0.1
Werkzeug==2.0.2
gunicorn
python-dotenv